sorry for the long commit message I relaise this would have been easier with mutiple commits with the different parts of my code, like I would do in proffesional environment. I havnt coded using pheonix before so I ended up breaking alot of things and just making everything work was an accomplishment in itself.
objectives:
frontend:
-added transaction modal to add transactions
-added a transaction table to show the transactions
-added bulma css library to give the site style and life
-added a UI and generated several graphql hooks to query and mutate the database
-added a currency rate functionality to allow user's to convert the currency in the transactions table. this uses an API to pull the latest currency convertion rates.
backend:
-added a method to seed the database useing seeds.ex file, however it only seeds companies, users, and merchants the UI should be used to create transactions
-added the companies schema
-added amount type in ./homework/transactions/amount to allow users to enter a transaction with decimals.

bonus:
-fixed the bug in transactions to allow users to add credit and update to credit. I resolved this bug by adding credit to the transactions changeset
-fixed; this may or may not be a bug but in elixir route you have it set up to accept requests from '/graphiql' but on the frontend apollo client you have it setup to
 request from '/graphql'

Notes:
elixir and graphql were fairly new to me so I may have done some things without using best practices or done in the hardest way. I did this project in elixir vs Node to show
that I'm a quick learner and not my experience with elixier, so keep that in mind. One last item I spent quite a bit of time on this project but can only spend so much time. So
there are alot of functionality missing from this project mainly error catching that I would always have on a production site, just keep that in mind. Other than that enjoy and
feel free to ask me question regarding this project when we chat.